Optimize performance. (#27)

+++ test/runtests.jl
@@ -18,7 +18,7 @@ crs = EPSG(4326)
         idx = index(pointcloud, cellsizes; crs=crs)
         index(collect(pointcloud), cellsizes; crs=crs, bbox=GeoInterface.extent(pointcloud))
         @inferred index(pointcloud, cellsizes; crs=crs)
-        @test typeof(idx) == PointCloudRasterizers.PointCloudIndex{LazIO.Dataset{0x00},Int}
+        @test typeof(idx) == PointCloudRasterizers.PointCloudIndex{LazIO.Dataset{0x00},UInt16}
         @test parent(idx) === pointcloud
         @test maximum(counts(idx)) == 2
         @test counts(idx).f.linear[1] == 1.0
